Rename Organization to the plural name
