+++ App/src/App.svelte
@@ -2,7 +2,7 @@
     import Head from './document/Head.svelte'
     import Body from './document/Body.svelte'
     import Sign from './document/layout/Sign.svelte'
-    import Organization from './document/layout/Organization.svelte'
+    import Organizations from './document/layout/Organizations.svelte'
 
     import LogoHalf from '$cmp/logo/LogoHalf.svelte'
     import ListPCC from '$cmp/lists/ListPCC.svelte'
@@ -23,5 +23,5 @@
     <ListPCC/>
     <ModalAlert/>
 {:else}
-    <Organization/>
+    <Organizations/>
 {/if}
